Mark all children of a failed commit as "failed"
Because a git history is modeled as a graph, a single commit cannot be
judged in isolation. When one commit fails the CLA check, none of its
children should be labeled as "passing". This is especially important in
the GitHub Pull Request UI because it reports the state of a given Pull
Request as the state of the final commit.