Fix #135 by ignoring web committers
GitHub specifies the commiter as being "web-flow" with an e-mail of NULL if you make a commit using the GitHub website. However the contributors information is properly included in the author section. Since copyright is with the author we've decided to ignore checking the web committer for CLA Hub purposes.