Stop saving signature OAuth tokens
If a user who is a maintainer of a repo later is signed out from CLAHub and signs in from a signature page their stored OAuth token is overwritten with one that has lesser permissions. This changes the behavior so we do not store the OAuth token from signatories since we should only need it for that session.

Resolves #134
Resolves #144